# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorField"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorField": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ErrorField",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MedicationDetailResponse": {
        "properties": {
          "active_ingredients": {
            "title": "Active Ingredients",
            "type": "string"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "classification": {
            "enum": [
              "OTC",
              "PRESCRIPTION"
            ],
            "title": "Classification",
            "type": "string"
          },
          "dosage_instructions": {
            "title": "Dosage Instructions",
            "type": "string"
          },
          "efficacy": {
            "title": "Efficacy",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "package": {
            "$ref": "#/components/schemas/MedicationPackageResponse"
          },
          "permit_code": {
            "title": "Permit Code",
            "type": "string"
          },
          "precautions": {
            "title": "Precautions",
            "type": "string"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/MedicationSourceResponse"
          },
          "storage_instructions": {
            "title": "Storage Instructions",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "brand",
          "name",
          "image_url",
          "package",
          "permit_code",
          "classification",
          "active_ingredients",
          "efficacy",
          "dosage_instructions",
          "precautions",
          "storage_instructions",
          "source"
        ],
        "title": "MedicationDetailResponse",
        "type": "object"
      },
      "MedicationListItemResponse": {
        "properties": {
          "active_ingredients": {
            "title": "Active Ingredients",
            "type": "string"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "classification": {
            "enum": [
              "OTC",
              "PRESCRIPTION"
            ],
            "title": "Classification",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "package": {
            "$ref": "#/components/schemas/MedicationPackageResponse"
          },
          "permit_code": {
            "title": "Permit Code",
            "type": "string"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "brand",
          "name",
          "image_url",
          "package",
          "permit_code",
          "classification",
          "active_ingredients"
        ],
        "title": "MedicationListItemResponse",
        "type": "object"
      },
      "MedicationListResponse": {
        "examples": [
          {
            "has_next": false,
            "items": [
              {
                "active_ingredients": "개발용 예시 성분 A",
                "brand": "영양꾹 로컬 테스트",
                "classification": "OTC",
                "id": "22000000-0000-4000-8000-000000000010",
                "image_url": "/static/products/local-medication-a.svg",
                "name": "복용 관리 예시 의약품 A",
                "package": {
                  "unit_form": "TABLET",
                  "units_per_package": "20"
                },
                "permit_code": "LOCAL-MED-001",
                "sku": "LOCAL-MED-001"
              }
            ],
            "page": 1,
            "page_size": 20,
            "total": 2
          }
        ],
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MedicationListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "has_next"
        ],
        "title": "MedicationListResponse",
        "type": "object"
      },
      "MedicationPackageResponse": {
        "properties": {
          "unit_form": {
            "enum": [
              "TABLET",
              "CAPSULE",
              "SCOOP",
              "PACKET"
            ],
            "title": "Unit Form",
            "type": "string"
          },
          "units_per_package": {
            "pattern": "^(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "title": "Units Per Package",
            "type": "string"
          }
        },
        "required": [
          "unit_form",
          "units_per_package"
        ],
        "title": "MedicationPackageResponse",
        "type": "object"
      },
      "MedicationSourceResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reviewed_on": {
            "format": "date",
            "title": "Reviewed On",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "reviewed_on"
        ],
        "title": "MedicationSourceResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessCookieAuth": {
        "description": "HttpOnly access JWT 쿠키",
        "in": "cookie",
        "name": "yeongyangkkuk_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "영양꾹 1차 로컬 MVP 백엔드 API",
    "title": "영양꾹 백엔드",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/medications": {
      "get": {
        "operationId": "medication_list_catalog",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationListResponse"
                }
              }
            },
            "description": "게시된 시드 의약품 페이지",
            "headers": {
              "Cache-Control": {
                "description": "인증 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "의약품 카탈로그 목록 조회",
        "tags": [
          "의약품"
        ]
      }
    },
    "/api/v1/medications/{product_id}": {
      "get": {
        "operationId": "medication_get_catalog_item",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationDetailResponse"
                }
              }
            },
            "description": "출처를 포함한 시드 의약품 상세",
            "headers": {
              "Cache-Control": {
                "description": "인증 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "MEDICATION_NOT_FOUND",
                    "fields": [],
                    "message": "의약품을 찾을 수 없습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "게시 의약품 상세 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "의약품 카탈로그 상세 조회",
        "tags": [
          "의약품"
        ]
      }
    },
...
  }
}